Create user management partial stage 3

+++ app/Http/Controllers/API/UserController.php
@@ -17,6 +17,7 @@ class UserController extends Controller
   public function __construct()
   {
       $this->middleware('auth:api');
+      // $this->authorize('isAdmin');
   }
 
     /**
@@ -69,9 +70,17 @@ class UserController extends Controller
         if($request->photo != $currentPhoto){
         $name = time().'.' . explode('/', explode(':', substr($request->photo, 0, strpos($request->photo, ';')))[1])[1];
         \Image::make($request->photo)->save(public_path('img/profile/').$name);
-
         $request->merge(['photo' => $name]);
+
+        $userPhoto = public_path('img/profile/').$currentPhoto;
+        if(file_exists($userPhoto)){
+          @unlink($userPhoto);
+        }
       }
+
+      if(!empty($request->password)){
+        $request->merge(['password' => Hash::make($request['password'])]);
+      };
 
       $user->update($request->all());
       return ['message' => "Success"];
@@ -120,12 +129,14 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
+      $this->authorize('isAdmin');
 
-        //delete users
+      $user = User::findOrFail($id);
 
-        $user->delete();
+      //delete users
 
-        return ['message' => 'User deleted'];
+      $user->delete();
+
+      return ['message' => 'User deleted'];
     }
 }
